requirements.txt
Django==3.0.5
django-debug-toolbar==2.2
psycopg2==2.8.5
django-filter==2.2.0
gunicorn
slackclient==2.5.0
django-allauth==0.41.0